refactor: 프로필 화면 UI 개선
